AWS/AG_DMV_Arize_Eval_Lambda_V2/requirements.txt
opentelemetry-api
opentelemetry-sdk
openinference-semantic-conventions
arize-otel
wrapt